requirements.txt
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==1.9
django-crispy-forms==1.7.2
django-formtools==1.0
django-import-export==1.0.0
django-reversion==1.8.5
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
future==0.16.0
get==0.0.39
httplib2==0.9.2
idna==2.6
jdcal==1.3
MySQL-python==1.2.5
odfpy==1.3.6
openpyxl==2.5.0
Pillow==5.0.0
post==0.0.26
public==0.0.65
PyYAML==3.12
query-string==0.0.28
request==0.0.26
requests==2.18.4
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
